+++ types/protocol-mapping.d.ts
@@ -562,13 +562,6 @@ export namespace ProtocolMapping {
          */
         'Storage.interestGroupAuctionEventOccurred': [Protocol.Storage.InterestGroupAuctionEventOccurredEvent];
         /**
-         * Specifies which auctions a particular network fetch may be related to, and
-         * in what role. Note that it is not ordered with respect to
-         * Network.requestWillBeSent (but will happen before loadingFinished
-         * loadingFailed).
-         */
-        'Storage.interestGroupAuctionNetworkRequestCreated': [Protocol.Storage.InterestGroupAuctionNetworkRequestCreatedEvent];
-        /**
          * Shared storage was accessed by the associated page.
          * The following parameters are included in all events.
          */
@@ -4292,8 +4285,7 @@ export namespace ProtocolMapping {
             returnType: void;
         };
         /**
-         * Enables/Disables issuing of interestGroupAuctionEventOccurred and
-         * interestGroupAuctionNetworkRequestCreated.
+         * Enables/Disables issuing of interestGroupAuctionEvent events.
          */
         'Storage.setInterestGroupAuctionTracking': {
             paramsType: [Protocol.Storage.SetInterestGroupAuctionTrackingRequest];
@@ -4850,6 +4842,10 @@ export namespace ProtocolMapping {
         };
         'FedCm.clickDialogButton': {
             paramsType: [Protocol.FedCm.ClickDialogButtonRequest];
+            returnType: void;
+        };
+        'FedCm.openUrl': {
+            paramsType: [Protocol.FedCm.OpenUrlRequest];
             returnType: void;
         };
         'FedCm.dismissDialog': {