+++ types/protocol.d.ts
@@ -15232,11 +15232,6 @@ export namespace Protocol {
         export type InterestGroupAuctionEventType = ('started' | 'configResolved');
 
         /**
-         * Enum of network fetches auctions can do.
-         */
-        export type InterestGroupAuctionFetchType = ('bidderJs' | 'bidderWasm' | 'sellerJs' | 'bidderTrustedSignals' | 'sellerTrustedSignals');
-
-        /**
          * Ad advertising element inside an interest group.
          */
         export interface InterestGroupAd {
@@ -15892,23 +15887,6 @@ export namespace Protocol {
              * Set for started and configResolved
              */
             auctionConfig?: any;
-        }
-
-        /**
-         * Specifies which auctions a particular network fetch may be related to, and
-         * in what role. Note that it is not ordered with respect to
-         * Network.requestWillBeSent (but will happen before loadingFinished
-         * loadingFailed).
-         */
-        export interface InterestGroupAuctionNetworkRequestCreatedEvent {
-            type: InterestGroupAuctionFetchType;
-            requestId: Network.RequestId;
-            /**
-             * This is the set of the auctions using the worklet that issued this
-             * request.  In the case of trusted signals, it's possible that only some of
-             * them actually care about the keys being queried.
-             */
-            auctions: InterestGroupAuctionId[];
         }
 
         /**
@@ -17978,6 +17956,11 @@ export namespace Protocol {
         export type DialogButton = ('ConfirmIdpLoginContinue' | 'ErrorGotIt' | 'ErrorMoreDetails');
 
         /**
+         * The URLs that each account has
+         */
+        export type AccountUrlType = ('TermsOfService' | 'PrivacyPolicy');
+
+        /**
          * Corresponds to IdentityRequestAccount
          */
         export interface Account {
@@ -18013,6 +17996,12 @@ export namespace Protocol {
         export interface ClickDialogButtonRequest {
             dialogId: string;
             dialogButton: DialogButton;
+        }
+
+        export interface OpenUrlRequest {
+            dialogId: string;
+            accountIndex: integer;
+            accountUrlType: AccountUrlType;
         }
 
         export interface DismissDialogRequest {